+++ src/main/java/ro/ubbcluj/apm/am/repository/doctor/CsvFileDoctorRepository.java
@@ -3,6 +3,7 @@ package ro.ubbcluj.apm.am.repository.doctor;
 import ro.ubbcluj.apm.am.config.ApplicationConfig;
 import ro.ubbcluj.apm.am.domain.Doctor;
 import ro.ubbcluj.apm.am.repository.FileRepository;
+import ro.ubbcluj.apm.am.util.FileUtil;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -11,13 +12,12 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.URL;
 import java.util.List;
 
 public class CsvFileDoctorRepository extends FileRepository<Integer, Doctor> {
 
     public CsvFileDoctorRepository() throws IOException {
-        this.filePath = ApplicationConfig.getInstance().getCsvFileDoctorRepositoryPath();
+        this.filePath = ApplicationConfig.getInstance().getDoctorRepositoryResourcePath("csv");
         readFromFile();
     }
 
@@ -46,12 +46,7 @@ public class CsvFileDoctorRepository extends FileRepository<Integer, Doctor> {
 
     @Override
     protected void writeToFile() throws IOException {
-        URL url = CsvFileDoctorRepository.class.getResource(this.filePath);
-        if (url == null) {
-            throw new FileNotFoundException("File not found: " + this.filePath);
-        }
-        String path = url.getPath();
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(path))) {
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(FileUtil.getFilePath(this.filePath)))) {
             writer.write("id,name,specialty,location,grade");
             writer.newLine();
             for (Doctor doctor : map.values()) {
@@ -66,37 +61,6 @@ public class CsvFileDoctorRepository extends FileRepository<Integer, Doctor> {
                 writer.write(String.valueOf(doctor.getGrade()));
                 writer.newLine();
             }
-        }
-    }
-
-    @Override
-    public Doctor add(Doctor doctor) {
-        Doctor addedDoctor = super.add(doctor);
-        try {
-            writeToFile();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        return addedDoctor;
-    }
-
-    @Override
-    public void update(Doctor value) {
-        super.update(value);
-        try {
-            writeToFile();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    public void deleteById(Integer id) {
-        super.deleteById(id);
-        try {
-            writeToFile();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
 }